requirements.txt
biopython
pandas
numpy
matplotlib